+++ js/src/main.js
@@ -1,109 +1,171 @@
-//small server meant to receive text/image files, handle directory structure,
-//pass them appropriately to the converter, and send them home to tilde
-const express = require('express'),
-	fileUpload = require('express-fileupload'),
-	app = express(),
-	{ exec } = require('child_process'),
-	config = require('./config.json'),
-	d = process.argv[2]//debug flag
-var res, soft
-app.use(fileUpload())
+const { exec } = require('child_process'),
+	config = require('../config.js'),
+	d = process.argv.includes('-d')//debug flag
+let localpath, res, soft, dry, test, cwd
 
-app.get('/', function(req, rs) {
-	rs.sendFile(__dirname + '/index.html')
-}).post('/upload', function(req, rs) {
+
+function upload(req, rs, path) {
+	debug(req.body)
+	for(let x in req.body) {//checkbox values are returned as 'boolean', so convert
+		req.body[x] = req.body[x] === 'true' ? true : (req.body[x] === 'false' ? false : req.body[x])
+	}
+	debug(req.body)
 	res = rs
 	if(!req.files)
-		return sendHead('no files were uploaded')
-	if(req.body.soft)
-		soft = true
-	let file = req.files.file//could be single object, or array of multiple files, so...
-	if(isIterable(file)) {//...we do this messy fix
-		debug('iterable')
-		for(let x of file)
-			handle(x, req.body)
+		return sendHead(400, 'no files were uploaded')
+
+	soft = req.body.soft
+	test = req.body.test
+	dry = test || req.body.dry
+
+	localpath = path
+	cwd = path + (test ? 'test/' : '')
+
+	if(req.body.edit) {
+		exists(req.body.editname, (result) => {
+			if(result)
+				store(req.files['0'], 'text/'+req.body.editname+'.txt', () => {
+					run('update', req.body.editname, req.body, () => {
+						sendHead(200, '')
+					})
+				})
+			else
+				sendHead(404, `no post ${req.body.editname} found`)
+		})
+
 	} else {
-		debug('single file')
-		handle(file, req.body)
+		debug(req.files)
+		let file = req.files//could be single object, or contain multiple, so...
+		if(isIterable(file)) {//...we do this messy fix
+			debug('iterable')
+			let runs = file.length, i = 0
+			for(let x of file)
+				handle(x, req.body, () => {
+					if(runs === ++i)
+						sendHead(200, '')
+				})
+		} else {
+			debug('single file')
+			handle(file['0'], req.body, () => {
+				sendHead(200, 'OK')
+			})
+		}
 	}
-	setTimeout(() => {
-		sendHead('', 1)
-	}, 10000)//THIS IS A REALLY BAD SOLUTION TO ASYNCHRONICITY BULLSHIT BUT YOU CAN'T BLAME ME
-}).post('/special', function(req, rs) {
+}
+
+function special(req, rs, path) {
+	localpath = path
 	res = rs
 	debug(req.body)
-	if(req.body.compile) {
-		debug('compiling')
-		exec(`${config.path}main compile`, (error, stdout, stderr) => {
-			if (error) {
-				sendHead(`during converter run, ${error.code + stderr + stdout}`)
-			}
-			debug(stdout + ' ' + stderr)
-			scp('index.html')//send out compiled index
-			sendHead('', 1)
-		})
-	} else {
-		sendHead('no parameter passed')
-	}
-})
+	switch(req.body.special) {
 
-function handle(file, body) {
+		case 'compile':
+			debug('compiling')
+			exec(`${localpath}main -a compile`, (error, stdout, stderr) => {
+				debug(stdout + ' ' + stderr)
+				if (error) {
+					sendHead(400, `during converter run: ${stderr}`)
+				} else {
+					scp('index.html', 'out', () => {
+						sendHead(200, '')
+					})//send out compiled index
+				}
+			})
+			break
+
+		case 'download':
+			exists(req.body.download, (result) =>  {
+				if(result)
+					res.sendFile(`${localpath}text/${req.body.download}.txt`)
+				else 
+					sendHead(404, `no post ${req.body.download} found`)
+			})
+			break
+
+		default:
+			sendHead(400, 'no parameter passed')
+	}
+}
+
+function handle(file, body, done) {
 	let match = file.mimetype.match(/text|image/)
 	if(!match)
-		return sendHead(`none or unsuccesful match against mimetype ${file.mimetype}`)
-	else {
-		let name = file.name.replace(' ', '-'),
-			loc = `${match[0]}/${name}`// local path + name
-		debug('handling '+name)
-		store(file, loc, () => {//asynchronicity make me want to be dead
-			if(match[0] == 'text') {
-				debug('exec')
-				let cmd = `${config.path}main add "${name}"${body.music?' "'+body.music+'" ':''}${body.musicurl?' "'+body.musicurl+'" ':''}`
-				debug(cmd)
-				if(!soft)
-					exec(cmd, (error, stdout, stderr) => {
-						if (error) {
-							sendHead(`during converter run, ${error.code + stderr + stdout}`)
-						}
-						console.log(stdout + ' ' + stderr)
-						scp('index.html')//send out compiled index
-					})
-			} else
-				scp(name, 1)//send out image
+		return sendHead(400, `none or unsuccesful match against mimetype ${file.mimetype}`)
+	let name = file.name.replace(' ', '-'),
+		loc = `${match[0]}/${name}`// local path + name
+	debug('handling '+name)
+	store(file, loc, () => {
+		if(match[0] === 'text')
+			run('add', name, body, () => {
+				done()
+			})
+		else
+			scp(name, match[0], () => {
+				done()
+			})//send out file
+	})
+
+}
+
+function run(action, name, body, done) {
+	debug('exec')
+	let cmd = `${localpath}main -a ${action} -f ${name}${body.music?` -m=${body.music}`:''}${body.musicurl?` -murl=${body.musicurl}`:''}`
+	debug(cmd)
+	if(!soft)
+		exec(cmd, {cwd:cwd}, (error, stdout, stderr) => {
+			if (error) {
+				sendHead(400, `during converter run: ${stderr}`)
+			}
+			console.log(stdout + ' ' + stderr)
+			scp('index.html', 'out', () => {
+				done()
+			})//send out compiled index
 		})
-	}
+	else
+		done()
 }
 
 function store(file, loc, done) {//write incoming file
 	if(soft)
 		done()
-	else {
-		file.mv(config.path+loc, function(err) {
-			if (err)
-				return res.status(500).send(err)
-			debug('stored')
-			done()
-		})
-	}
+	else
+		file.mv(cwd+loc)
+			.then(() => {
+				debug('stored')
+				done()
+			}).catch((err) => {
+				return sendHead(500, err.Error)
+			})
 }
 
-function scp(name, img) {//copy file to tilde
-	if(!soft)
-		exec(`scp ${config.path}${img?'image/':'out/'}${name} ${config.tilde.username}@tilde.town:${config.tilde.path}${img?'zenbu/img/':''}${name}`, (error, stdout, stderr) => {
+function scp(name, type, done) {//copy file to tilde
+	if(soft || dry)
+		done()
+	else
+		exec(`scp ${localpath}${type}/${name} ${config.tilde.username}@tilde.town:/home/${config.tilde.username}/${config.tilde.path}${config.tilde.path.endsWith('/')?'':'/'}${type!=='out'?`zenbu/${type}/`:''}${name}`, (error, stdout, stderr) => {
 			if (error) {
-				sendHead(`during scp, ${error}`)
+				sendHead(400, `during scp: ${error}`)
 			}
 			debug(error + ' ' + stdout + ' ' + stderr)
+			done()
 		})
 }
 
-function sendHead(string, ok) {
+function exists(name, done) {
+	exec(`${localpath}main -a exists -f ${name}`, {cwd:cwd}, (error, stdout, stderr) => {
+		debug(error + ' ' + stdout + ' ' + stderr)
+		if (error)
+			done(false)
+		else
+			done(true)
+	})
+}
+
+function sendHead(code, string) {
+	debug(string)
 	if(res.headersSent)
 		return
-	if(ok)
-		res.status(200).end()
-	else
-		res.status(400).send(`Something went wrong${string ? ': '+string : ''}`)
+	res.status(code).send(string)
 }
 
 function isIterable(obj) {
@@ -119,5 +181,11 @@ function debug(string) {
 		console.log(string)
 }
 
-app.listen(config.port)
-debug(`Server running at port ${config.port}`)
+module.exports = {
+	pages: [
+		'upload',
+		'special'
+	],
+	upload: upload,
+	special: special
+}